requirements.txt
uvicorn==0.44.0
fastapi==0.136.0
pydantic==2.13.3
jinja2==3.1.6
starlette==0.46.0
click==8.3.2
h11==0.16.0
colorama==0.4.6
pydantic-core==2.46.3
anyio==4.13.0
typing-extensions==4.15.0
typing-inspection==0.4.2
annotated-doc==0.0.4
annotated-types==0.7.0
MarkupSafe==3.0.3
idna==3.11
requests==2.33.1
charset-normalizer==3.4.7
urllib3==2.6.3
certifi==2026.2.25
python-dotenv==1.2.2
